requirements.txt
easy-rouge==0.2.2
gensim==3.7.2
matplotlib==3.0.2
nltk==3.5
numpy==1.17.2
pandas==0.23.4
py-readability-metrics==1.4.3
scikit-learn==0.22.2.post1
scipy==1.2.1
seaborn==0.9.0
tqdm==4.28.1
